New types for HTTPBody
